+++ Pandora/Pattern/Transformer.hs
@@ -2,3 +2,4 @@ module Pandora.Pattern.Transformer (module Exports) where
 
 import Pandora.Pattern.Transformer.Lowerable as Exports
 import Pandora.Pattern.Transformer.Liftable as Exports
+import Pandora.Pattern.Transformer.Hoistable as Exports