+++ Pandora/Pattern/Transformer/Hoistable.hs
@@ -1,0 +1,6 @@
+module Pandora.Pattern.Transformer.Hoistable (Hoistable (..)) where
+
+import Pandora.Core.Functor (type (~>))
+
+class Hoistable t where
+	hoist :: u ~> v -> t u ~> t v