+++ CHANGELOG.md
@@ -174,3 +174,4 @@
 * Define `Semiring` typeclass
 * Define `Stream` module and type
 * Move `Liftable` and `Lowerable` modules from `Functor` to `Transformer` submodule
+* Define `Hoistable` module and typeclass